The 1st prototype of individual works (update p5.js file))
I made some changes in this prototype based on the group code because I chose audio animation.

1. Added 2 variables (song, analyzer); these variables are added at the top of the script to hold the sound file and the amplitude analyzer, respectively.

2. Added preload function: load a new song suitable for my theme.

3. Make changes to the setup function:
for example, an amplitude analyzer, which Initializes a new amplitude analyzer object and sets the analyser's input to the loaded song to measure its amplitude.

4. Button for play/pause: through the mouse press can play or pause the song.

5. Modified draw function:
add rms to retrieve the current amplitude level of the song, and draw a rectangle at the center of the canvas with its size influenced by the amplitude of the song, creating a visual representation of the sound's volume.

6. Added play/pause function.