The 2nd prototype of individual works (update p5.js file)
The following are the changed I made based on the 1st prototype.

1.	Add more variables (song, fft, numBins, smoothing and button) to control the rectangles and audio track.

2.	setScale in display() method to set the factor for the rectangle can change in scale

3.	Set up function:
Create a new instance of p5.ftt() object and connect it to the song.

4.	Resize the button let it fit the window size through change.

5.	In draw function:
Analyzer() method has been added, to return an array of amplitude values (0-255) across the spectrum value, and use array to control the amplitude and frequency of rectangles through the sound has play and pause.

Set specValue as spectrum value and use ‘i’ divided by numbins to ensure the index ‘i’ wraps around within the range [0, numBins-1], and will not out of range.

Calculates the scale for the rectangles based on the spectrum value next.